Constrain the tilt view to the camera preview area.
On an iPhone, the camera preview doesn't take up the full screen. This more-precise sizing lets the grid lines also serve as composition lines, dividing the frame into thirds.

On an iPad, the camera preview does take up the full screen, but this causes a problem with gestures. The "Use Photo" / "Retake" buttons can't receive touch events if the tilt view is covering them. Previously, this was hacked by reducing the size of the tilt view by 20%. Now, we explicitly deny hit tests, which passes the touches down the Z stack to the built-in controls.