Hide the tilt view immediately after the user takes a photo, and show it again if the user taps "Retake."
The initial implementation of this commit had a considerable lag in hiding and reshowing the tilt view on the iPhone 6, so the following optimizations were added:

While the tilt view is hidden, there's no need to keep processing device motion updates. Stop the updates, and resume them once the tilt view becomes visible again.

Additionally, there is no need to look at anything other than the latest motion reading. Any intermediate readings are irrelevant, so cancel them out of the queue.

Finally, assign a low priority to the main queue operations that call setNeedsDisplay(). This allows operations that set isHidden to jump ahead. If the tilt view is hidden, we can no-op the setNeedsDisplay() operations when they eventually come in.