On iPad, don't let the tilt view draw on top of the controls.
The camera preview on iPad takes up the full screen, and the controls float on top. Since the tilt view is sized to the camera preview, that causes a collision.

We need to mask the tilt view, rather than resizing it, in order to keep the proportions correct. The tilt view grid lines should be at the 1/3 mark used for traditional photographic composition.