Bug fix: the gridlines sometimes disappear on iPad.
The mask that prevents the gridlines from drawing over the camera controls sometimes fails to resize correctly. Interestingly, the bug's repro appears to be correlated with the specific version of Xcode used to compile the app, not with the specific device or version of iPadOS that the app runs on.

The fix is to resize the mask explicitly.