New physics model!
"Pitch" is always drawn on the screen's Y axis (Home button to camera, on an iPhone 6 - SE 3).
"Roll" is always drawn on the screen's X axis (Volume buttons to power button, on an iPhone 6 - SE 3).

This eliminates the need to transform the attitude to accommodate UI rotation or lack thereof.

This should fix certain drawing bugs when the device is flat (face up or face down), or on the iPad when the UI orientation is locked.